requirements.txt
requests
schedule
psycopg2-binary
pytz
discord.py
